+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Changed
+* Nested `subtype` allowed
 
 ## [2.0.1](https://pypi.org/project/multimethod/2.0.1/) - 2025-11-02
 ### Changed